Version 1.3
Made the colorscheme more complete.  Now the status line is highlighted differently for the current buffer, and I changed a few other highlighting this to round out the colorscheme.  Also added a function to toggle the semicolon highlighting feature off and on with an nmap.  Very handy for c/c++, java, etc...